recode authcontroller and refactor entire controller
